Rename validation classes to package for consistent with entity
