+++ src/main/java/com/a_smart_cookie/controller/command/guest/RegistrationCommand.java
@@ -9,7 +9,7 @@ import com.a_smart_cookie.dto.sign_up.UserSignUpDto;
 import com.a_smart_cookie.entity.User;
 import com.a_smart_cookie.service.ServiceFactory;
 import com.a_smart_cookie.service.UserService;
-import com.a_smart_cookie.util.validation.UserValidator;
+import com.a_smart_cookie.util.validation.user.UserValidator;
 import org.apache.log4j.Logger;
 
 import javax.servlet.http.HttpServletRequest;
