+++ src/main/java/com/a_smart_cookie/util/validation/user/UserValidator.java
@@ -1,4 +1,4 @@
-package com.a_smart_cookie.util.validation;
+package com.a_smart_cookie.util.validation.user;
 
 import com.a_smart_cookie.dao.EntityColumn;
 import com.a_smart_cookie.dto.sign_up.UserSignUpDto;
